Moved errors into their own module.
Breaking change: slack::Error is no longer publicly exported. Use slack::error::Error instead.
